Create archives using git-archive
This change is motivated by issues in the Zip and Tarwriter
implementations causing the the same Git repository at the same revision
to produce archives with different checksums depending on the time you
create the archive.

This implementation is based on the Grit::Git::Repo#archive_to_file
method we added to the gitlab-grit gem in the past.